pyproject.toml
[tool.poetry]
name = "calculatepsd"
version = "0.1.0"
description = ""
authors = ["perdanamain <firman.fp123@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"
numpy = "^2.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
